refactor: extract invariant helpers into a separate file for better organization and reuse
