+++ pkg/compiler/internal/binding_compiler.go
@@ -67,6 +67,7 @@ func (c *BindingCompiler) CompileVariableDeclaration(ctx fql.IVariableDeclaratio
 	mutable := c.isMutableDeclaration(ctx)
 	storage := c.declarationStorage(decl, mutable)
 	src := c.exprs.Compile(ctx.Expression())
+
 	if src == bytecode.NoopOperand {
 		return bytecode.NoopOperand
 	}
