+++ pkg/compiler/internal/core/invariant.go
@@ -1,7 +1,5 @@
 package core
 
-import "fmt"
-
 // InvariantViolation marks a broken compiler-internal assumption.
 // These panics should indicate bugs in compiler state management rather than user input errors.
 type InvariantViolation struct {
@@ -10,14 +8,4 @@ type InvariantViolation struct {
 
 func (e *InvariantViolation) Error() string {
 	return e.message
-}
-
-func PanicInvariant(message string) {
-	panic(&InvariantViolation{
-		message: "compiler invariant violated: " + message,
-	})
-}
-
-func PanicInvariantf(format string, args ...any) {
-	PanicInvariant(fmt.Sprintf(format, args...))
 }
