+++ pkg/compiler/internal/core/invariant_helpers.go
@@ -1,0 +1,13 @@
+package core
+
+import "fmt"
+
+func PanicInvariant(message string) {
+	panic(&InvariantViolation{
+		message: "compiler invariant violated: " + message,
+	})
+}
+
+func PanicInvariantf(format string, args ...any) {
+	PanicInvariant(fmt.Sprintf(format, args...))
+}
